refactor(openpi): unify pytorch conversion
