+++ examples/embodiment/config/model/pi0_5_rlinf.yaml
@@ -4,8 +4,7 @@ model_type: "openpi_rlinf"
 # eval experiment config (behavior_openpi_pi05_rlinf_eval.yaml); the values here
 # are placeholders so this template stays path-free and reusable.
 
-# Weight dtype for eval. Dual-Franka deployment defaults to full precision.
-precision: fp32
+precision: bf16
 
 # LoRA is not used for eval, but the shared model factory reads these keys.
 is_lora: False
